feat(audit): dedicated detail page per audit at /dashboard/audit/[id]
Split the Site Audit page into a hub (run bar + recent audits) and a
per-audit detail route. Each audit is now a bookmarkable/shareable URL;
running an audit navigates to its detail page, which polls until complete.
Recent audits link to their detail pages. The report UI is extracted into a
shared AuditReport component. Web-only; no API/schema changes.
